subgraphs/users/requirements.txt
anyio==4.9.0
click==8.1.8
dnspython==2.7.0
email_validator >=2.2.0
fastapi==0.115.12
graphql-core==3.2.6
greenlet==3.1.1
h11==0.16.0
idna==3.10
libcst==1.7.0
msgpack==1.1.0
mypy-extensions==1.0.0
pathspec==0.12.1
phonenumbers==9.0.2
platformdirs==4.3.7
pydantic>=2.11.2,<3.0.0
pydoc-markdown==4.8.2
Pygments==2.19.1
pymongo==4.13.2
python-ldap==3.4.4
python-multipart>=0.0.20
requests>=2.32.3
rich==14.0.0
six==1.17.0
sniffio==1.3.1
starlette>=0.41.2,<0.42.0
strawberry-graphql[fastapi]==0.263.2
tomli==2.2.1
typer==0.15.2
typing-extensions>=4.13.1
uvicorn==0.34.0
